requirements.txt
fastapi==0.135.3
uvicorn==0.43.0
jinja2==3.1.6
python-multipart==0.0.22
httpx==0.28.1
aiofiles==25.1.0
watchdog==6.0.0
python-dotenv==1.2.2
